requirements.txt
Django==4.2.7
gunicorn==21.2.0
whitenoise==6.6.0
pymysql==1.1.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
python-decouple==3.8
dj-database-url==2.3.0

